created separate class file; updated code.
